refactor(nx-adsp): move the MCP server wiring out of the quality module
utils/quality.ts is per-project quality tooling — six of its exports take a
projectRoot or projectName and configure ESLint, Jest coverage, Semgrep, axe,
or Playwright for one project. addAdspMcpServer takes only the Tree, writes
workspace-root files, and after the previous commit was also the only function
in the module mutating package.json, with a dependency version constant sitting
above the ESLint rules. That's the wrong module.

utils/agent.ts is a false friend — it's the ADSP agent-service consultation, a
runtime conversation with a platform agent, not coding-agent tooling. So this
follows the utils/nginx.ts precedent instead: one module per concern, the
single place to edit to change that output everywhere.

Behavior is unchanged and coverage is unaffected — the MCP wiring is tested
through the generators that call it (init.spec.ts, express-service.spec.ts),
and quality.spec.ts never covered it.
